chore: moves html reconstruction on client side
